+++ backend/Api/Models/DatabaseModels.cs
@@ -4,6 +4,7 @@ namespace Api.Models
     {
         public int Id { get; set; }
         public required string Username { get; set; }
+        public required string Name { get; set; }
         public required string Email { get; set; }
         public DateTime CreatedAt { get; set; }
