+++ backend/Api/Models/JwtSettings.cs
@@ -1,0 +1,10 @@
+namespace Api.Models
+{
+    public class JwtSettings
+    {
+        public string SecretKey { get; set; } = "";
+        public string Issuer { get; set; } = "";
+        public string Audience { get; set; } = "";
+        public int ExpiryMinutes { get; set; } = 60;
+    }
+}
